+++ aftiktuna-rust/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Idle animation for all creatures
 - New indoor decorations
 - Json location definition `crew_ship.json` for the ship and new symbol types "ship_controls" and "food_deposit"
+- Noise message when something happens in a nearby area
 
 ### Changed